Create README.md
# Edmonton-Wards-Statistics
I took a picture of all the of the wards in Edmonton. From that picture I utilized my photoshop skills to highlight each Ward. 

From data I found on the internet I compiled and averaged data based on each ward. Through this I was able to find the average household income in each neighborhood, the population and the average income tax in each ward. 

This data is to hover every time you go to the certain ward through the use of javascript. The new photo is supposed to show up through javascript is as well. 

I choose this project due to its practicality in the real world. This is a tool that can be utilized easily and has information that people are interested in. I made it easier for people to view a million inputs from the consensus in an interactive format.